app_front/pyproject.toml
[project]
name = "app-front"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair<5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.19.0",
]
